Added `progress_callback`

+++ bot/__init__.py
@@ -9,21 +9,25 @@ from pytgcalls import GroupCallFactory
 dir = os.path.dirname(os.path.abspath(__file__))
 folder = os.path.join(dir, "logs")
 
-
-if os.path.exists(f"{folder}/logs.txt"):
-    pass
-elif os.stat(f"{folder}/logs.txt").st_size > 0:
+if not os.path.exists(folder):
+    os.makedirs(folder)
+if (
+    os.path.exists(f"{folder}/logs.txt")
+    and os.stat(f"{folder}/logs.txt").st_size > 0
+):
     with open(f"{folder}/logs.txt", "w") as f:
         f.truncate(0)
+        print(
+            f'{colored("[INFO]", "blue")}: LOG FILE WAS FLUSHED SUCCESSFULLY'
+        )
         f.close()
-elif not os.path.exists(folder):
-    os.makedirs(folder)
-try:
-    with open(f"{folder}/logs.txt", "w") as f:
-        f.write("")
-    print(f'{colored("[INFO]", "blue")}: LOG FILE CREATED')
-except Exception as e:
-    print(e)
+elif not os.path.exists(f"{folder}/logs.txt"):
+    try:
+        with open(f"{folder}/logs.txt", "w") as f:
+            f.write("")
+        print(f'{colored("[INFO]", "blue")}: LOG FILE CREATED')
+    except Exception as e:
+        logging.err(e)
 
 
 logging.basicConfig(
